Added Teleswift GUI wrapper
Introducing: the Teleswift GUI wrapper. Write your bot code and implement it in a GUI with a nice logging screen and various controls that are built for Teleswift.

Uses the Console class' built-in GUI functionality to log to a NSTextView.

Warning: This wrapper may contain code that may have been written hastily/poorly. It's not great, but it works.

Harry